FISH
A fish (pl.: fish or fishes) is an aquatic, anamniotic, gill-bearing vertebrate animal with swimming fins and a hard skull, but lacking limbs with digits. Fish can be grouped into the more basal jawless fish and the more common jawed fish, the latter including all living cartilaginous and bony fish, as well as the extinct placoderms and acanthodians. In a break to the long tradition of grouping all fish into a single class (Pisces), contemporary phylogenetics views fish as a paraphyletic group.

Most fish are cold-blooded, their body temperature varying with the surrounding water, though some large active swimmers like white shark and tuna can hold a higher core temperature. Many fish can communicate acoustically with each other, such as during courtship displays. The study of fish is known as ichthyology.

The earliest fish appeared during the Cambrian as small filter feeders; they continued to evolve through the Paleozoic, diversifying into many forms. The earliest fish with dedicated respiratory gills and paired fins, the ostracoderms, had heavy bony plates that served as protective exoskeletons against invertebrate predators. The first fish with jaws, the placoderms, appeared in the Silurian and greatly diversified during the Devonian, the "Age of Fishes".

Bony fish, distinguished by the presence of swim bladders and later ossified endoskeletons, emerged as the dominant group of fish after the end-Devonian extinction wiped out the apex predators, the placoderms. Bony fish are further divided into the lobe-finned and ray-finned fish. About 96% of all living fish species today are teleosts, a crown group of ray-finned fish that can protrude their jaws. The tetrapods, a mostly terrestrial clade of vertebrates that have dominated the top trophic levels in both aquatic and terrestrial ecosystems since the Late Paleozoic, evolved from lobe-finned fish during the Carboniferous, developing air-breathing lungs homologous to swim bladders. Despite the cladistic lineage, tetrapods are usually not considered to be fish.

Fish have been an important natural resource for humans since prehistoric times, especially as food. Commercial and subsistence fishers harvest fish in wild fisheries or farm them in ponds or in breeding cages in the ocean. Fish are caught for recreation, or raised by fishkeepers as ornaments for private and public exhibition in aquaria and garden ponds. Fish have had a role in human culture through the ages, serving as deities, religious symbols, and as the subjects of art, books and movies.